backend/requirements.txt
fastapi[standard]
pydantic
uvicorn
opencv-python
numpy
starlette
pycryptodome